Finish making 3 new components
